+++ test/helper-float-test.js
@@ -8,7 +8,7 @@ var options = {}
 
 describe('Parser Helpers', function () {
   describe('helpers.float()', function () {
-    false &&describe('with an invalid float #1', function () {
+    describe('with  an invalid float #1', function () {
       var Parser = atokParser.createParserFromFile('./parsers/float2HelperParser.js', 'options')
       var p = new Parser(options), err
 
@@ -16,19 +16,19 @@ describe('Parser Helpers', function () {
         function handler (token, idx, type) {
           switch (type) {
             case 'data':
+              done()
             break
             default:
-              err = new Error('handler should not be called')
+              done( new Error('handler should not be called') )
           }
         }
 
         p.on('data', handler)
         p.write('- ')
-        done(err)
       })
     })
 
-    false &&describe('with an invalid float #1', function () {
+    false&&describe('with  an invalid float #2', function () {
       var Parser = atokParser.createParserFromFile('./parsers/float2HelperParser.js', 'options')
       var p = new Parser(options), err
 
@@ -56,6 +56,7 @@ describe('Parser Helpers', function () {
         function handler (token, idx, type) {
           switch (type) {
             case 'float':
+              assert.equal(typeof token, 'number')
               assert.equal(token, 123)
             break
             default:
@@ -69,7 +70,7 @@ describe('Parser Helpers', function () {
       })
     })
 
-    describe('with a full float as a negative integer', function () {
+    describe('with  a full float as a negative integer', function () {
       var Parser = atokParser.createParserFromFile('./parsers/floatHelperParser.js', 'options')
       var p = new Parser(options), err
 
@@ -77,6 +78,7 @@ describe('Parser Helpers', function () {
         function handler (token, idx, type) {
           switch (type) {
             case 'float':
+              assert.equal(typeof token, 'number')
               assert.equal(token, -123)
             break
             default:
@@ -90,7 +92,7 @@ describe('Parser Helpers', function () {
       })
     })
 
-    describe('with a full float as a positive float', function () {
+    describe('with  a full float as a positive float', function () {
       var Parser = atokParser.createParserFromFile('./parsers/floatHelperParser.js', 'options')
       var p = new Parser(options), err
 
@@ -98,6 +100,7 @@ describe('Parser Helpers', function () {
         function handler (token, idx, type) {
           switch (type) {
             case 'float':
+              assert.equal(typeof token, 'number')
               assert.equal(token, 123.456)
             break
             default:
@@ -111,7 +114,7 @@ describe('Parser Helpers', function () {
       })
     })
 
-    describe('with a full float as a negative float', function () {
+    describe('with  a full float as a negative float', function () {
       var Parser = atokParser.createParserFromFile('./parsers/floatHelperParser.js', 'options')
       var p = new Parser(options), err
 
@@ -119,6 +122,7 @@ describe('Parser Helpers', function () {
         function handler (token, idx, type) {
           switch (type) {
             case 'float':
+              assert.equal(typeof token, 'number')
               assert.equal(token, -123.456)
             break
             default:
@@ -132,7 +136,7 @@ describe('Parser Helpers', function () {
       })
     })
 
-    describe('with a full float as a positive float with exponent', function () {
+    describe('with  a full float as a positive float with exponent', function () {
       var Parser = atokParser.createParserFromFile('./parsers/floatHelperParser.js', 'options')
       var p = new Parser(options), err
 
@@ -140,6 +144,7 @@ describe('Parser Helpers', function () {
         function handler (token, idx, type) {
           switch (type) {
             case 'float':
+              assert.equal(typeof token, 'number')
               assert.equal(token, 123.456e7)
             break
             default:
@@ -153,7 +158,7 @@ describe('Parser Helpers', function () {
       })
     })
 
-    describe('with a full float as a negative float with exponent', function () {
+    describe('with  a full float as a negative float with exponent', function () {
       var Parser = atokParser.createParserFromFile('./parsers/floatHelperParser.js', 'options')
       var p = new Parser(options), err
 
@@ -161,6 +166,7 @@ describe('Parser Helpers', function () {
         function handler (token, idx, type) {
           switch (type) {
             case 'float':
+              assert.equal(typeof token, 'number')
               assert.equal(token, -123.456e7)
             break
             default:
@@ -174,7 +180,7 @@ describe('Parser Helpers', function () {
       })
     })
 
-    describe('with an integer with exponent', function () {
+    describe('with  an integer with exponent', function () {
       var Parser = atokParser.createParserFromFile('./parsers/floatHelperParser.js', 'options')
       var p = new Parser(options), err
 
@@ -182,6 +188,7 @@ describe('Parser Helpers', function () {
         function handler (token, idx, type) {
           switch (type) {
             case 'float':
+              assert.equal(typeof token, 'number')
               assert.equal(token, 123e7)
             break
             default:
@@ -195,7 +202,7 @@ describe('Parser Helpers', function () {
       })
     })
 
-    describe('with a negative integer with exponent', function () {
+    describe('with  a negative integer with exponent', function () {
       var Parser = atokParser.createParserFromFile('./parsers/floatHelperParser.js', 'options')
       var p = new Parser(options), err
 
@@ -203,6 +210,7 @@ describe('Parser Helpers', function () {
         function handler (token, idx, type) {
           switch (type) {
             case 'float':
+              assert.equal(typeof token, 'number')
               assert.equal(token, -123e7)
             break
             default:
@@ -216,7 +224,7 @@ describe('Parser Helpers', function () {
       })
     })
 
-    describe('with a float with a negative exponent', function () {
+    describe('with  a float with a negative exponent', function () {
       var Parser = atokParser.createParserFromFile('./parsers/floatHelperParser.js', 'options')
       var p = new Parser(options), err
 
@@ -224,6 +232,7 @@ describe('Parser Helpers', function () {
         function handler (token, idx, type) {
           switch (type) {
             case 'float':
+              assert.equal(typeof token, 'number')
               assert.equal(token, 123e-7)
             break
             default:
@@ -237,7 +246,7 @@ describe('Parser Helpers', function () {
       })
     })
 
-    describe('with a float with a positive exponent', function () {
+    describe('with  a float with a positive exponent', function () {
       var Parser = atokParser.createParserFromFile('./parsers/floatHelperParser.js', 'options')
       var p = new Parser(options), err
 
@@ -245,6 +254,7 @@ describe('Parser Helpers', function () {
         function handler (token, idx, type) {
           switch (type) {
             case 'float':
+              assert.equal(typeof token, 'number')
               assert.equal(token, 123e7)
             break
             default:
@@ -258,7 +268,7 @@ describe('Parser Helpers', function () {
       })
     })
 
-    describe('with a split up float', function () {
+    describe('with  a split up float', function () {
       var Parser = atokParser.createParserFromFile('./parsers/floatHelperParser.js', 'options')
       var p = new Parser(options), err
 
@@ -266,6 +276,7 @@ describe('Parser Helpers', function () {
         function handler (token, idx, type) {
           switch (type) {
             case 'float':
+              assert.equal(typeof token, 'number')
               assert.equal(token, 123.456)
             break
             default:
@@ -280,7 +291,7 @@ describe('Parser Helpers', function () {
       })
     })
 
-    describe('with a split up float 2', function () {
+    describe('with  a split up float 2', function () {
       var Parser = atokParser.createParserFromFile('./parsers/floatHelperParser.js', 'options')
       var p = new Parser(options), err
 
@@ -288,6 +299,7 @@ describe('Parser Helpers', function () {
         function handler (token, idx, type) {
           switch (type) {
             case 'float':
+              assert.equal(typeof token, 'number')
               assert.equal(token, 123.456)
             break
             default:
@@ -302,7 +314,7 @@ describe('Parser Helpers', function () {
       })
     })
 
-    describe('with a split up float 3', function () {
+    describe('with  a split up float 3', function () {
       var Parser = atokParser.createParserFromFile('./parsers/floatHelperParser.js', 'options')
       var p = new Parser(options), err
 
@@ -310,6 +322,7 @@ describe('Parser Helpers', function () {
         function handler (token, idx, type) {
           switch (type) {
             case 'float':
+              assert.equal(typeof token, 'number')
               assert.equal(token, 123.456e7)
             break
             default:
@@ -325,7 +338,7 @@ describe('Parser Helpers', function () {
       })
     })
 
-    describe('with a split up float 4', function () {
+    describe('with  a split up float 4', function () {
       var Parser = atokParser.createParserFromFile('./parsers/floatHelperParser.js', 'options')
       var p = new Parser(options), err
 
@@ -333,6 +346,7 @@ describe('Parser Helpers', function () {
         function handler (token, idx, type) {
           switch (type) {
             case 'float':
+              assert.equal(typeof token, 'number')
               assert.equal(token, 123.456e7)
             break
             default:
@@ -348,7 +362,7 @@ describe('Parser Helpers', function () {
       })
     })
 
-    describe('with a split up float 5', function () {
+    describe('with  a split up float 5', function () {
       var Parser = atokParser.createParserFromFile('./parsers/floatHelperParser.js', 'options')
       var p = new Parser(options), err
 
@@ -356,6 +370,7 @@ describe('Parser Helpers', function () {
         function handler (token, idx, type) {
           switch (type) {
             case 'float':
+              assert.equal(typeof token, 'number')
               assert.equal(token, 123.456e17)
             break
             default:
@@ -371,7 +386,7 @@ describe('Parser Helpers', function () {
       })
     })
 
-    describe('with a split up float 6', function () {
+    describe('with  a split up float 6', function () {
       var Parser = atokParser.createParserFromFile('./parsers/floatHelperParser.js', 'options')
       var p = new Parser(options), err
 
@@ -379,6 +394,7 @@ describe('Parser Helpers', function () {
         function handler (token, idx, type) {
           switch (type) {
             case 'float':
+              assert.equal(typeof token, 'number')
               assert.equal(token, 123.456)
             break
             default:
@@ -393,7 +409,7 @@ describe('Parser Helpers', function () {
       })
     })
 
-    describe('with a non ending number', function () {
+    describe('with  a non ending number', function () {
       var Parser = atokParser.createParserFromFile('./parsers/floatHelperParser.js', 'options')
       var p = new Parser(options), err
       
@@ -403,6 +419,7 @@ describe('Parser Helpers', function () {
         function handler (token, idx, type) {
           switch (type) {
             case 'float':
+              assert.equal(typeof token, 'number')
               assert.equal(token, '123')
               res = token
             break